+++ CHANGELOG.rst
@@ -10,7 +10,7 @@ Upon version 0.8, and before version 1, SV2 major version increments are reflect
 Changelog
 =========
 
-* Add support for CIF topologies in libmda using simtk.openmm.app.pdbxfile
+* Add support for CIF topologies in libmda using openmm.app.pdbxfile
 
 v0.10.0 (2021-11-24)
 ------------------------------------------------------------
